MeetingList 타입 전부 수정 및 전환
